[type: refactor]refactor logging plugin code
